everything ready to start writing model tests
