+++ Rakefile
@@ -1,0 +1,6 @@
+require 'rake/testtask'
+
+Rake::TestTask.new do |t|
+  t.libs << "test"
+  t.pattern = "tests/**/*_test.rb"
+end
