+++ tests/models/skill_inventory_test.rb
@@ -1,0 +1,20 @@
+require_relative "../test_helper.rb"
+
+class SkillInventoryTest < Minitest::Test
+
+  def test_all_returns_all_skills
+  end
+
+  def test_add_makes_a_new_skill
+  end
+
+  def test_find_finds_a_skill
+  end
+
+  def test_can_update_a_skill
+  end
+
+  def test_can_delete_a_skill
+  end
+
+end
